优化object_op的单测
