+++ unittest/src/object_op_retry_test.cpp
@@ -18,7 +18,7 @@ class ObjectOpRetryTest : public ::testing::Test {
       static void SetUpTestCase() {
         m_client_without_retry= createCosClient(0);
         m_client_with_retry= createCosClient(3);
-        m_bucket_name = GetEnvVar("ErrBucket") + "-" + GetEnvVar("ErrAppid");
+        m_bucket_name = "coscppsdkv5ut-retry-" + GetEnvVar("ErrBucket") + "-" + GetEnvVar("ErrAppid");
       }
 
       static CosAPI* createCosClient(int max_retry_num) {
@@ -50,7 +50,7 @@ CosAPI* ObjectOpRetryTest::m_client_with_retry;
 std::string ObjectOpRetryTest::m_bucket_name;
 
 TEST_F(ObjectOpRetryTest, GetObjectFor2xxTest) {
-    {
+  {
     qcloud_cos::GetObjectByFileReq req(m_bucket_name, "200", "./test.txt");
     qcloud_cos::GetObjectByFileResp resp;
